requirements.txt
ply==3.3
SPARQLWrapper==1.5.1
pydot==1.0.28
